Add Risy DAO Token (RISY) on Polygon Mainnet (#50)
### Description
Risy DAO Token is the core cryptocurrency of the Risy ecosystem used for
votes, dApps, HODL, and more. Risy DAO is a revolutionary decentralized
autonomous organization built on the Polygon Mainnet blockchain for fast
and affordable dApps. We're addressing critical issues in the
cryptocurrency space, including centralization, market manipulation, and
unfair token distribution. Our mission is to create a truly
decentralized, community-driven ecosystem that continuously evolves to
meet the needs of its users. With Risy DAO we build a more equitable,
stable, democratic, and sustainable financial future, especially by
preventing whale manipulations. Risy DAO Token's liquidity is 50% USDC
and 50% PAXG (which is burned to avoid rug pulls) so It differs from
other cryptocurrencies by correlating with gold and the dollar, rather
than with other cryptocurrencies such as Bitcoin.

**Important Note:** Risy DAO will not attempt to list on centralized
exchanges to ensure that Risy DAO Token, a truly decentralized
cryptocurrency, remains decentralized, not correlated with BTC or any
other cryptocurrency, and not monopolized by any centralized exchange.
Instead, it will provide initiatives and education to encourage people's
crypto literacy and adaptation to the decentralized world.

**Therefore, we encourage tools to display the price of the Risy DAO
Token from DEX built-in oracles.**

### Risy DAO Token
* **Network:** Polygon Mainnet
* **Token Address:** **`0xca154cF88F6ffBC23E16B5D08a9Bf4851FB97199`**
* **Symbol:** RISY
* **Polygonscan (verified):**
https://polygonscan.com/token/0xca154cF88F6ffBC23E16B5D08a9Bf4851FB97199

#### Liquidities
* **CoinMarketCap**

> * **RISY/USDC:**
https://coinmarketcap.com/dexscan/polygon/0xb908228a001cb177ac785659505ebca1d9947ee8/
> * **RISY/PAXG:**
https://coinmarketcap.com/dexscan/polygon/0x8341b5240e05552d85e78bcd691b2982c3e4deaf/

* **GeckoTerminal**

> * **RISY/USDC:**
https://www.geckoterminal.com/polygon_pos/pools/0xb908228a001cb177ac785659505ebca1d9947ee8
> * **RISY/PAXG:**
https://www.geckoterminal.com/polygon_pos/pools/0x8341b5240e05552d85e78bcd691b2982c3e4deaf

* **Proof of Burned Liquidity Ownership**

> * **RISY/USDC:**
https://polygonscan.com/tx/0xf54e5f5b1402b8a7e2e4dbdbd413e3bb620bcb594b188d927a5052628f6caa9b
> * **RISY/PAXG:**
https://polygonscan.com/tx/0x29a3fa3d33acddb416c6f9bbc138b99ab703b4d63bc8df2e4aff2b2072acb4f1

### Official Website, Important Links and Socials
* **Website:** https://risy.io
* **Official DAO:** https://www.tally.xyz/gov/risydao
* **Whitepaper:** https://risy.io/files/Risy%20DAO%20Whitepaper.pdf
* **X:** https://x.com/RisyDAO
* **Telegram Community:** https://t.me/RisyDAOCommunity
* **Discord Community:** https://discord.gg/zuSBq2XBjc
* **Source Code:** https://github.com/RisyDAO
* **Medium News & Blog:** https://risydao.medium.com **(Learn more about
Risy DAO on Medium)**
* **Trade on QuickSwap:**
https://quickswap.exchange/#/swap?currency0=ETH&swapIndex=1&currency1=0xca154cF88F6ffBC23E16B5D08a9Bf4851FB97199